requirements.txt
absl-py==0.12.0
affine==2.3.0
alabaster==0.7.12
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
attrs==21.1.0
Babel==2.9.0
bce-python-sdk==0.8.59
cachetools==4.2.2
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
cycler==0.10.0
Cython==0.29.22
decorator==4.4.2
distlib==0.3.1
docutils==0.16
filelock==3.0.12
flake8==3.9.0
Flask==1.1.2
Flask-Babel==2.0.0
future==0.18.2
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.37.1
h5py==2.10.0
identify==2.2.2
idna==2.10
image-similarity-measures==0.3.5
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==3.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
mkl-fft==1.3.0
mkl-service==2.3.0
networkx==2.5.1
nodeenv==1.5.0
nose==1.3.7
numpy==1.20.3
oauthlib==3.1.0
opencv-python==4.3.0.36
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
paddlepaddle-gpu==2.0.1
pandas==1.1.3
phasepack==1.5
Pillow==8.2.0
pre-commit==2.11.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycallgraph==1.0.1
pycocotools==2.0.2
pycodestyle==2.7.0
pycryptodome==3.10.1
pyFFTW==0.12.0
pyflakes==2.3.1
Pygments==2.8.1
pykalman==0.9.5
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
python-vlc==3.0.11115
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
rasterio==1.2.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.23.2
scipy==1.6.2
seaborn==0.11.0
shellcheck-py==0.7.1.1
sklearn==0.0
snowballstemmer==2.1.0
snuggs==1.4.7
Sphinx==3.5.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.4.8
toml==0.10.2
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.3
visualdl==2.1.1
Werkzeug==1.0.1
wrapt==1.12.1
yappi==1.3.2
yml==0.0.1
zipp==3.4.1
